requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
openai==2.31.0
python-dotenv==1.0.0
pypdf==6.10.0
